Updated db script

+++ inv_tooldb.sql
@@ -31,7 +31,7 @@ DROP TABLE IF EXISTS `kit_components`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `kit_components` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `kit_id` int(11) NOT NULL,
   `component_id` int(11) NOT NULL,
   `quantity` int(11) NOT NULL,
@@ -172,4 +172,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2019-07-26  8:54:02
+-- Dump completed on 2019-07-26  9:09:03
